feat: Add members table to database
This commit adds the necessary code changes to create the members table in the database. The members table will store information about the members of the music platform. The changes include:
- Adding the members table to the database.sql file.
- Defining the columns for the members table, including id, name, and description.

Note: The commit messages "Merge branch 'main' of https://github.com/caiquenaimi/Final-project-backend" and "git init: first commit" are not relevant to the current code changes and can be ignored.