feat: Add birthdate column to members table
This commit adds the birthdate column to the members table in the database. The birthdate column is required and will store the birthdate of each member. This change ensures that the members table captures all necessary information about the members of the music platform.

Note: The commit messages "Merge branch 'main' of https://github.com/caiquenaimi/Final-project-backend" and "git init: first commit" are not relevant to the current code changes and can be ignored.