feat: Remove album and artist controllers and routes
This commit removes the albumController, artistsController, albumRoutes, and artistsRoutes files. These files are no longer needed as the functionality they provided has been refactored and integrated into other parts of the codebase. This change helps to simplify the code structure and improve maintainability.

Note: The commit messages "Merge branch 'main' of https://github.com/caiquenaimi/Final-project-backend" and "git init: first commit" are not relevant to the current code changes and can be ignored.