feat: Add login route for users
This commit adds a new login route for users in the usersRoutes.js file. The route is defined as a POST request to "/users/login" and is handled by the loginUser function in the usersController.js file. This allows users to log in to the application using their email and password. The loginUser function queries the database to find a user with the provided email and password, and returns a JSON response with the user's information if a match is found. If no user is found or the password is incorrect, an appropriate error message is returned. This feature enhances the functionality of the application by providing a secure login mechanism for users.