feat: Add rtoken table to database
This commit adds a new table called "rtoken" to the database. The "rtoken" table has columns for "rtoken" (VARCHAR), "expires" (INT), and "user_id" (INT). The "user_id" column is a foreign key referencing the "id" column in the "users" table. This change enhances the functionality of the application by allowing the storage of refresh tokens for user authentication.