feat: Update password hashing to use bcrypt
This commit updates the password hashing mechanism in the usersController.js file to use the bcrypt npm package. By replacing the previous hash function with bcrypt.hash, we improve the security of user passwords by using a stronger and more secure hashing algorithm. This enhancement ensures that user passwords are properly protected in the application.