fix: Correct logout route in usersRoutes.js
This commit fixes a typo in the logout route of the usersRoutes.js file. The route was previously defined as "/users/logout" but has been corrected to "/users/logOut" to match the corresponding function name in the usersController. This fix ensures that the logout functionality works as intended in the application.