feat: Update error message for invalid token
The commit modifies the `refreshToken` function in the `usersController.js` file to update the error message for an invalid token. Previously, the error message stated "Token inválido ou expirado" (Invalid or expired token), but it has been changed to "Token inválido" (Invalid token) to provide a more accurate description of the issue. This update improves the clarity of error messages in the application.