feat: Add endpoint to get music by ID
This commit adds a new endpoint to the `musicsRoutes.js` file that allows retrieving a music by its ID. The `getMusicById` function is implemented in the `musicsController.js` file to query the database and return the music if found. This feature enhances the functionality of the application by providing a way to fetch specific music details.