feat: Add updateMusicFavoriteStatus endpoint
This commit adds a new endpoint to the `musicsRoutes.js` file that allows updating the favorite status of a music record. The `updateMusicFavoriteStatus` function is implemented in the `musicsController.js` file to update the `favorite` field of a music record in the database based on the provided `id` parameter and `favorite` value in the request body. This feature enhances the functionality of the application by allowing users to mark their favorite music tracks.