feat: Update updateUser function to handle password updates
This commit updates the `updateUser` function in the `usersController.js` file to handle password updates. If a new password is provided in the request body, it will be hashed using bcrypt and updated in the database. This enhancement improves the security of the application by ensuring that user passwords are properly encrypted and stored.